refactor: separate folders and links operations
